chore(react): refactor application generator for clarity
