+++ packages/react/src/generators/application/lib/add-linting.ts
@@ -1,0 +1,69 @@
+import {
+  type Tree,
+  type GeneratorCallback,
+  joinPathFragments,
+} from '@nx/devkit';
+import { Linter, lintProjectGenerator } from '@nx/eslint';
+import {
+  addExtendsToLintConfig,
+  addOverrideToLintConfig,
+  addPredefinedConfigToFlatLintConfig,
+  isEslintConfigSupported,
+} from '@nx/eslint/src/generators/utils/eslint-file';
+import { useFlatConfig } from '@nx/eslint/src/utils/flat-config';
+import { addDependenciesToPackageJson, runTasksInSerial } from '@nx/devkit';
+import { addSwcDependencies } from '@nx/js/src/utils/swc/add-swc-dependencies';
+import { extraEslintDependencies } from '../../../utils/lint';
+import { NormalizedSchema } from '../schema';
+
+export async function addLinting(host: Tree, options: NormalizedSchema) {
+  const tasks: GeneratorCallback[] = [];
+  if (options.linter === Linter.EsLint) {
+    const lintTask = await lintProjectGenerator(host, {
+      linter: options.linter,
+      project: options.projectName,
+      tsConfigPaths: [
+        joinPathFragments(options.appProjectRoot, 'tsconfig.app.json'),
+      ],
+      unitTestRunner: options.unitTestRunner,
+      skipFormat: true,
+      rootProject: options.rootProject,
+      skipPackageJson: options.skipPackageJson,
+      addPlugin: options.addPlugin,
+    });
+    tasks.push(lintTask);
+
+    if (isEslintConfigSupported(host)) {
+      if (useFlatConfig(host)) {
+        addPredefinedConfigToFlatLintConfig(
+          host,
+          options.appProjectRoot,
+          'flat/react'
+        );
+        // Add an empty rules object to users know how to add/override rules
+        addOverrideToLintConfig(host, options.appProjectRoot, {
+          files: ['*.ts', '*.tsx', '*.js', '*.jsx'],
+          rules: {},
+        });
+      } else {
+        const addExtendsTask = addExtendsToLintConfig(
+          host,
+          options.appProjectRoot,
+          { name: 'plugin:@nx/react', needCompatFixup: true }
+        );
+        tasks.push(addExtendsTask);
+      }
+    }
+
+    if (!options.skipPackageJson) {
+      const installTask = addDependenciesToPackageJson(
+        host,
+        extraEslintDependencies.dependencies,
+        extraEslintDependencies.devDependencies
+      );
+      const addSwcTask = addSwcDependencies(host);
+      tasks.push(installTask, addSwcTask);
+    }
+  }
+  return runTasksInSerial(...tasks);
+}
